模板 ci 支持 packages 仓库 ci
