+++ .github/workflows/pkgs-action.yml
@@ -4,43 +4,54 @@ on:
   workflow_call
 
 jobs:
-  test:
+  packages-test:
     name: ${{ github.repository }}
     runs-on: ubuntu-latest
     steps:
+      - uses: actions/checkout@v3
       - name: Install Tools
         shell: bash 
         run: |
           sudo apt install python3 python3-pip gcc git libncurses5-dev -y
           python3 -m pip install scons==4.4.0 requests tqdm wget html-table
-          url=https://raw.githubusercontent.com/RT-Thread/env/master/touch_env.sh
-          wget $url -O touch_env.sh
-          chmod 777 touch_env.sh
-          ./touch_env.sh
-          ~/.env/tools/scripts/pkgs --upgrade-modules
       - name: Install Test Resources
         shell: bash 
         run: |
-          cd ${{ github.workspace }} && git clone https://github.com/RT-Thread/pkgs-test.git
-          cd pkgs-test && python -c '__import__("'pkgs-test'").Config().get_resources();'
-      - name: Packages Test
+          cd ${{ github.workspace }}
+          PYTHONCMD="__import__('pkgs-test').Config().get_resources();"
+          if [[ ${{ github.repository }} == *'packages'* ]]; \
+          then cd .. && mkdir packages_temp && cp packages -R packages_temp && mv packages_temp packages/packages && cd packages; \
+          PYTHONCMD="import shutil; import os; config = __import__('pkgs-test').Config(); \
+          shutil.move('../packages', os.path.join(config.get_path('env'),'packages'));\
+          config.get_resources();"; fi
+          git clone https://github.com/RT-Thread/pkgs-test.git
+          cd pkgs-test && python -c "$PYTHONCMD"
+      - name: Carry Out Packages Test
         shell: bash 
         run: |
           cd ${{ github.workspace }}/pkgs-test
-          python -c 'import os; pkgs_test = __import__("'pkgs-test'"); config = pkgs_test.Config();\
+          python -c "import os; pkgs_test = __import__('pkgs-test'); config = pkgs_test.Config();\
           config.get_resources();\
-          packages_index = pkgs_test.PackagesIndex(os.path.join(os.environ["'HOME'"], "'.env/packages/packages'"));\
-          pkgs_config_dict = packages_index.repository_seek("'${{ github.repository }}'");\
-          logs = pkgs_test.Logs("'artifacts_export'", config.get_config_data(), pkgs_config_dict);\
-          build = pkgs_test.Build(config, pkgs_config_dict, logs); build.all();'
+          packages_index = pkgs_test.PackagesIndex(os.path.join(config.get_path('env'),'packages/packages'));\
+          pkgs_config_dict = packages_index.repository_seek('${{ github.repository }}');\
+          logs = pkgs_test.Logs('artifacts_export', config.get_config_data(), pkgs_config_dict);\
+          build = pkgs_test.Build(config, pkgs_config_dict, logs); build.all();"
       - uses: actions/upload-artifact@v3
         with:
           name: artifacts_export
           path: ${{ github.workspace }}/pkgs-test/artifacts_export
-      - name: Packages test error
+          
+  check-errors:
+    runs-on: ubuntu-latest
+    needs: packages-test
+    steps:
+      - name: Download artifacts_export
+        uses: actions/download-artifact@v3
+        with:
+          name: artifacts_export
+      - name: Packages test whether or not error
         shell: bash 
         run: |
-          cd ${{ github.workspace }}/pkgs-test/artifacts_export || true
           if [ -f "index.html" ]; then str1=`cat index.html`; else str1=""; fi
           echo $str1
           str2="#f00"
